+++ backend/mora/service/address.py
@@ -193,6 +193,7 @@ API
 
 '''
 
+import collections
 import json
 import re
 
@@ -230,11 +231,50 @@ HREF_PREFIXES = {
 }
 
 DEFAULT_ERROR = 'Fejl'
+NOT_FOUND = 'Ukendt'
 
 MUNICIPALITY_CODE_PATTERN = re.compile('urn:dk:kommune:(\d+)')
 
 blueprint = flask.Blueprint('address', __name__, static_url_path='',
                             url_prefix='/service')
+
+
+def _address_string_chunks(addr):
+    # loosely inspired by 'adressebetegnelse' in apiSpecification/util.js from
+    # https://github.com/DanmarksAdresser/Dawa/
+
+    yield addr['vejnavn']
+
+    if addr.get('husnr') is not None:
+        yield ' '
+        yield addr['husnr']
+
+    if addr.get('etage') is not None or addr.get('dør') is not None:
+        yield ','
+
+    if addr.get('etage') is not None:
+        yield ' '
+        yield addr['etage']
+        yield '.'
+
+    if addr.get('dør') is not None:
+        yield ' '
+        yield addr['dør']
+
+    yield ', '
+
+    if addr.get('supplerendebynavn') is not None:
+        yield addr['supplerendebynavn']
+        yield ', '
+
+    yield addr['postnr']
+    yield ' '
+    yield addr['postnrnavn']
+
+
+def stringify(addr):
+    '''Return a string representation of the given DAWA address object'''
+    return ''.join(_address_string_chunks(addr))
 
 
 def get_relation_for(addrobj, fallback=None):
@@ -312,57 +352,66 @@ def get_one_address(c, addrrel, class_cache=None):
     scope = util.checked_get(addrclass, 'scope', 'DAR')
 
     if scope == 'DAR':
-        def make_error_object(errordesc):
+        def make_error_object(errordesc, name=DEFAULT_ERROR):
+            flask.current_app.logger.warn(
+                'ADDRESS LOOKUP FAILED in {!r}:\n{}'.format(
+                    flask.request.url,
+                    errordesc,
+                ),
+            )
+
             return {
                 mapping.ADDRESS_TYPE: addrclass,
                 mapping.HREF: None,
-                mapping.NAME: DEFAULT_ERROR,
+                mapping.NAME: name,
                 mapping.UUID: addrrel['uuid'],
                 mapping.ERROR: errordesc,
             }
 
-        # unfortunately, we cannot live with struktur=mini, as it omits
-        # the formatted address :(
-        try:
-            r = session.get(
-                'https://dawa.aws.dk/adresser/' + addrrel['uuid'],
-                params={
-                    'noformat': '1',
-                },
-            )
-        except Exception as exc:
-            # the exception above is overly broad for a) safety and b)
-            # testing -- specifically, the exception raised by
-            # requests_mock does not descend from RequestException :(
-            util.log_exception('failed to get address ' + addrrel['uuid'])
+        for addrtype in ('adresser', 'adgangsadresser',
+                         'historik/adresser', 'historik/adgangsadresser'):
+            try:
+                r = session.get(
+                    'https://dawa.aws.dk/' + addrtype,
+                    # use a list to work around unordered dicts in Python < 3.6
+                    params=[
+                        ('id', addrrel['uuid']),
+                        ('noformat', '1'),
+                        ('struktur', 'mini'),
+                    ],
+                )
 
-            return make_error_object(str(exc))
+                addrobjs = r.json()
 
-        if not r.ok:
-            error = r.json()
+            except Exception as exc:
+                # the exception above is overly broad for a) safety and b)
+                # testing -- specifically, the exception raised by
+                # requests_mock does not descend from RequestException :(
+                return make_error_object(str(exc))
 
-            flask.current_app.logger.warn(
-                'ADDRESS LOOKUP FAILED in {!r}:\n{}'.format(
-                    flask.request.url,
-                    json.dumps(error, indent=2),
-                ),
-            )
+            if not r.ok:
+                return make_error_object(addrobjs)
 
-            return make_error_object(error)
+            if addrobjs:
+                # found, escape loop!
+                break
 
-        addrobj = r.json()
+        else:
+            return make_error_object(NOT_FOUND, NOT_FOUND)
+
+        addrobj = addrobjs.pop()
 
         return {
             mapping.ADDRESS_TYPE: addrclass,
             mapping.HREF: (
+                # link to the location on OSM; historical addresses
+                # may lack coordinates
                 'https://www.openstreetmap.org/'
-                '?mlon={}&mlat={}&zoom=16'.format(
-                    *addrobj['adgangsadresse']
-                    ['adgangspunkt']['koordinater']
-                )
+                '?mlon={x}&mlat={y}&zoom=16'.format(**addrobj)
+                if 'x' in addrobj and 'y' in addrobj else None
             ),
 
-            mapping.NAME: addrobj['adressebetegnelse'],
+            mapping.NAME: stringify(addrobj),
             mapping.UUID: addrrel['uuid'],
         }
 
@@ -660,7 +709,33 @@ def address_autocomplete(orgid):
     else:
         code = None
 
-    addrs = requests.get(
+    #
+    # In order to allow reading both access & regular addresses, we
+    # autocomplete both into an ordered dictionary, with the textual
+    # representation as keys. Regular addresses tend to be less
+    # relevant than access addresses, so we list them last.
+    #
+    # The limits are somewhat arbitrary: Since access addresses mostly
+    # differ by street number or similar, we only show five -- by
+    # comparison, ten addresses seems apt since they may refer to
+    # apartments etc.
+    #
+
+    addrs = collections.OrderedDict(
+        (addr['tekst'], addr['adgangsadresse']['id'])
+        for addr in session.get(
+            'https://dawa.aws.dk/adgangsadresser/autocomplete',
+            # use a list to work around unordered dicts in Python < 3.6
+            params=[
+                ('per_side', '5'),
+                ('noformat', '1'),
+                ('kommunekode', code),
+                ('q', q),
+            ],
+        ).json()
+    )
+
+    for addr in session.get(
         'https://dawa.aws.dk/adresser/autocomplete',
         # use a list to work around unordered dicts in Python < 3.6
         params=[
@@ -669,14 +744,15 @@ def address_autocomplete(orgid):
             ('kommunekode', code),
             ('q', q),
         ],
-    ).json()
+    ).json():
+        addrs.setdefault(addr['tekst'], addr['adresse']['id'])
 
     return flask.jsonify([
         {
             "location": {
-                "uuid": addr['adresse']['id'],
-                "name": addr['tekst']
-            }
+                "name": k,
+                "uuid": addrs[k],
+            },
         }
-        for addr in addrs
+        for k in addrs
     ])